Refactor folder unit test
